requirements.txt
# ============================================
# E-Commerce Fashion Image Generation
# Requirements for RTX 4060 (8GB VRAM)
# ============================================

# --------------------------------------------
# PyTorch with CUDA 12.1 support
# --------------------------------------------
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.2.0+cu121
torchvision==0.17.0+cu121
torchaudio==2.2.0+cu121

# --------------------------------------------
# Basic utilities (needed for data scripts)
# --------------------------------------------
tqdm==4.66.1                    # Progress bars
Pillow==10.2.0                  # Image processing
numpy==1.26.3
pyyaml==6.0.1                   # Config files

# --------------------------------------------
# Projected GAN Dependencies
# --------------------------------------------
timm==0.9.12                    # EfficientNet and other pretrained models
clean-fid==0.1.35               # FID score calculation
ninja==1.11.1.1                 # For custom CUDA kernels
lpips==0.1.4                    # Perceptual loss
scipy==1.11.4                   # Scientific computing
click==8.1.7                    # CLI interface

# --------------------------------------------
# Stable Diffusion + LoRA Dependencies
# --------------------------------------------
diffusers==0.25.1               # Hugging Face diffusion models
transformers==4.37.2            # Tokenizers and text encoders
accelerate==0.26.1              # Distributed training utilities
peft==0.8.2                     # Parameter-Efficient Fine-Tuning (LoRA)
safetensors==0.4.2              # Safe model serialization

# --------------------------------------------
# Data Handling & Image Processing
# --------------------------------------------
datasets==2.17.0                # Hugging Face datasets
albumentations==1.3.1           # Image augmentations
opencv-python==4.9.0.80         # Computer vision utilities
gdown==5.1.0                    # Google Drive downloads

# --------------------------------------------
# Training & Logging
# --------------------------------------------
tensorboard==2.15.1             # Training visualization
wandb==0.16.3                   # Weights & Biases logging (optional)
einops==0.7.0                   # Tensor operations

# --------------------------------------------
# Utilities
# --------------------------------------------
matplotlib==3.8.2
imageio==2.33.1
imageio-ffmpeg==0.4.9           # For video generation
rich==13.7.0                    # Beautiful terminal output

# --------------------------------------------
# xformers - MUST be installed separately after PyTorch
# See install instructions below
# --------------------------------------------
# NOTE: xformers is optional but recommended for 8GB VRAM
# Install manually after other packages:
# pip install xformers --index-url https://download.pytorch.org/whl/cu121
